requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
gunicorn==20.0.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
django-cors-middleware==1.5.0
djangorestframework==3.11.0
django-cors-headers==3.2.1
Markdown==3.2
django-filter==2.2.0
flake8==3.7.9
django-related-admin==0.7.0
mysql-client